requirements.txt
pytdx
pandas
numpy
requests
python-dotenv
